box doesn;t move :(

+++ src/Movable.java
@@ -1,10 +1,10 @@
 public class Movable {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public double direction;
     public double speed;
 
-    public Movable (int x, int y, double speed){
+    public Movable (float x, float y, double speed){
         this.x = x;
         this.y = y;
         this.speed = speed;
